feat: handle LinkRTT packets
Handle LinkRTT packets on in_links. Without this the in_link rtt was always 0.

The reference implementation has an "initiator" flag in the Link structure. This seems to correspond
to out_links in this implementation. Added an out_link flag to handle_packet and handle_data_packet
to indicate whether the packet should be handled as initiator or not.